+++ packages/expo-cli/src/commands/credentials.ts
@@ -1,44 +1,18 @@
-import { CommanderStatic } from 'commander';
+import type { Command } from 'commander';
 
-import { Context, runCredentialsManagerStandalone } from '../credentials';
-import {
-  SelectAndroidExperience,
-  SelectIosExperience,
-  SelectPlatform,
-} from '../credentials/views/Select';
+import { applyAsyncActionProjectDir } from './utils/applyAsyncAction';
 
-type Options = {
-  platform?: 'android' | 'ios';
-  parent?: {
-    nonInteractive: boolean;
-  };
-};
-
-export default function (program: CommanderStatic) {
-  program
-    .command('credentials:manager [path]')
-    .description('Manage your credentials')
-    .helpGroup('credentials')
-    .option('-p --platform <platform>', 'Platform: [android|ios]', /^(android|ios)$/i)
-    .asyncActionProjectDir(
-      async (projectRoot: string, options: Options) => {
-        const context = new Context();
-        await context.init(projectRoot, {
-          nonInteractive: options.parent?.nonInteractive,
-        });
-        let mainpage;
-        if (options.platform === 'android') {
-          mainpage = new SelectAndroidExperience();
-        } else if (options.platform === 'ios') {
-          mainpage = new SelectIosExperience();
-        } else {
-          mainpage = new SelectPlatform();
-        }
-        await runCredentialsManagerStandalone(context, mainpage);
-      },
-      {
-        checkConfig: false,
-        skipSDKVersionRequirement: true,
-      }
-    );
+export default function (program: Command) {
+  applyAsyncActionProjectDir(
+    program
+      .command('credentials:manager [path]')
+      .description('Manage your credentials')
+      .helpGroup('credentials')
+      .option('-p --platform <platform>', 'Platform: [android|ios]', /^(android|ios)$/i),
+    () => import('./credentialsManagerAsync'),
+    {
+      checkConfig: false,
+      skipSDKVersionRequirement: true,
+    }
+  );
 }
